Let build_images try to set path to bmpblk_utility
The user is supposed to provide path to bmpblk_utility - the tool
generated by the vboot_reference package. If the tool was generated in
the standard place, we should be able to find it, thus not requiring
the user to set up the PATH manually.

BUG=none
TEST=manual
   . tried the following commands in the top level directory

   $ make strings
   $ cd images
   $ make

    observed them succeed despite not defining path to bmpblk_utility
    in the environment.

Change-Id: Ibbfccc6c40a066f9f6cb697e6282d9145de6b624
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/39561
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Hung-Te Lin <hungte@google.com>