bmpblk: Support new board configuration "phy_rec".
Device with physical recovery button (without built-in keyboard) needs to enter
developer mode by recovery button instead of "Enter". That's a new property to
the firmware screens and we need to support that by the new "phy_rec"
configuration value in boards.yaml.

Note we don't have official translations yet, so the locales (except en) are
made by replacing "ENTER" in "todev" message to "RECOVERY" as a temporary
solution. Since no shipping boards are using the new message, this should be
fine; and we will fix when the final version of localization strings are ready.

BUG=chrome-os-partner:24963
TEST=make # Old boards all works fine
     Manually add a "phy_rec: True" to board 'std' and build std,
     found the correct new todev message in output bitmaps.

Change-Id: If8db006096b1513d35b2b83d595e83d779e6324a
Reviewed-on: https://chrome-internal-review.googlesource.com/151959
Tested-by: Hung-Te Lin <hungte@google.com>
Reviewed-by: Mohammed Habibulla <moch@google.com>
Commit-Queue: Mohammed Habibulla <moch@google.com>