split model line into left|model|right parts
Currently, the model line consists of two images: model.bmp, which
represents 'model' in each language and an actual model name. Right-
to-left languages require 'model' to be on the right hand side of
the model name.

This patch makes the model line consist of three images: model_left.bmp,
the model, and model_right.bmp. This way localization team can
handle what's the appropriate position of the text 'model' and we
don't need to have this knowledge in the code.

BUG=chromium:502066
BRANCH=tot
TEST=Tested on Jerry with/without USE=bitmap_in_cbfs

Change-Id: I4dd624b1f3dbb44479ce71a93ea88fe9b235ee31
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chrome-internal-review.googlesource.com/241482
Commit-Ready: Daisuke Nojiri <dnojiri@google.com>
Tested-by: Daisuke Nojiri <dnojiri@google.com>
Reviewed-by: Hung-Te Lin <hungte@google.com>