vboot: Show BadUSB image in the center
When the device doesn't support SD card recovery, BadSD.bmp is created as
an empty image. When Depthcharge renders a screen, it shows BadSD.bmp and
BadUSB.bmp side by side, resulting in USB stick image shifted to the right.

This change fixes the issue by creating BadDevices.bmp, which can be either
BadDevices.png or BadUSB.png, depending on the board configuration. This is
consistent with how we create InsertDevices.bmp.

BUG=chromium:501060
BRANCH=tot
TEST=Tested on Jerry in 50 languages with and without SD card support

Change-Id: I6e82be0ffe5119607355610f8d3184515dcd11b6
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/337599
Reviewed-by: Hung-Te Lin <hungte@chromium.org>