locale/en: Make detachable UI strings more consistent
This patch changes a few of the new detachable UI strings to be more
consistent with existing terminology. For the most part, we standardize
on the term "OS verification" that has always been used on existing
firmware screens (as opposed to "Root Verification", "Verified Boot",
"Developer Mode") to make it easier for our users and translators that
might not know that all of these mean the same thing.

Remove the "(not working yet)" from network boot since it's not super
useful... as long as it doesn't work, we should disable the whole menu
option to not even show the string. Later when we get it working, this
will save us another translation round-trip.

This patch only changes the English strings. We'll use it as a
blue-print for the translation team to update everything else.

BRANCH=None
BUG=b:38347807
TEST=None

Change-Id: Ia7555a91b6bcaee7f1a94829e1d047e7f7a4199a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/612517
Reviewed-by: Shelley Chen <shchen@chromium.org>