locale: Strip trailing newlines from single-field files
The translation files that only contain a single field to be translated
(language.txt and update.txt) currently don't strip trailing new-lines.
This is inconsistent with how the other files are handled and makes it
easy to accidentally include a trailing new-line in language.txt when
updating or adding a language, which makes it look bad on the screen.
Fix this to avoid future problems.

Also, don't parse the detachable strings file as a single big input
string in addition to the way it's meant to be parsed.

CQ-DEPEND=CL:639238
BRANCH=none
BUG=none
TEST=none

Change-Id: I68e4e1ab95c99d4a8074e2df0a53c2e7082ebc6f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/639018
Reviewed-by: Hung-Te Lin <hungte@chromium.org>