images: Update Chrome logo to increase resolution and circularity
The current Chrome logo (uploaded in CL:331831) is slightly oval for
some unclear reason. This patch replaces it with the image currently
used for the boot splash screen which is presumably officially
sanctioned. As an added bonus, this logo is not only much rounder but
also higher resolution, making sure it will continue to look crisp and
pretty on our secret upcoming 4K Chromebooks.

BRANCH=None
BUG=b:65479973
TEST=Booted Scarlet, compared semi-minor and semi-major axis of logo.

Change-Id: I7c44f748a6776807e59183624102eddaa15ea388
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903233
Reviewed-by: Hung-Te Lin <hungte@chromium.org>