Remove more unused images
This patch splits the existing firmware screens into those that are used
in all cases and those that are only needed by the non-detachable UI, so
that the latter ones don't need to be included on detachable builds. It
also removes a few extra strings and images that are no longer used from
images, so that we're not wasting flash space on junk we don't need.

Also get rid of some style flags that were essentially no-ops.

BRANCH=None
BUG=None
TEST=Built and looked at on Kevin and Scarlet.

Change-Id: I0b90f80ae081b514dd9392211e22841b4b68dd94
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903234
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>