strings: Fix "Model" string color style
The "Model" string was always meant to be styled in dark gray, same as
the "For help visit:" string. Apparently whoever put the style statement
in there either didn't quite understand how it works, or the handling of
the string was changed afterwards and the style not updated. Either way,
it doesn't work anymore and the string is incorrectly black. This patch
fixes the issue.

BRANCH=None
BUG=None
TEST=Booted Scarlet, confirmed color of the string.

Change-Id: I39f76b57bf0e2956ab622a394183263b9f013f39
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903236
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>