Grunt: Update screen resolution to 1920x1080
Increasing the resolution improves the quality of the recovery mode
screens which were very blurry at 1024x768 expanded to either 1366x768
or 1920x1080.  We found the correct video mode in the vBIOS for the
FHD screen, so we can now use the full size bitmaps.

BUG=b:111943807
TEST=Boot and test on 1920x1080 and 1366x768 screens

Change-Id: Ic11d1f4c176584398234525b4f3958cc8fbd58ac
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1297013
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>