bmpblk: add strings for diagnostic UI
This change adds a string/string for the diagnostic UI, and does so
in a way that shouldn't consume space on boards that don't use the
feature.

BUG=None
BRANCH=None
TEST=did a local compile and tested that string is present on board with
feature and not present on boards that don't use the feature.  Tested
that string is usable by the UI that utilizes the strings.

Change-Id: Ia5a9d0ed686da9ba3885f7b7e553654c0588c15f
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1509676
Reviewed-by: Julius Werner <jwerner@chromium.org>