bmpblk: fix mappings for building multiple boards
When using "make" in bmpblk I noticed that the mapping state from one
board was often carried to the next.  I hardly know python, but the
docs I read suggest that the mapping dictionary is declared
as a class member, so each instance is sharing the same 'replace_map'.
I've changed this so the existing (and renamed) class member is used
as a default for creating a per-instance mapping dictionary.

BUG=b:129471321
BRANCH=None
TEST=Ran "make" in bmpblk and noticed that the output about mappings
is now a proper reflection of what's present in images/boards.yaml.

Change-Id: I3b95c02657a0337fd8bbf9c9c979d2476e99ac82
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1581120
Commit-Ready: Matt Delco <delco@chromium.org>
Tested-by: Matt Delco <delco@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>