bmpblk: 3 different keys for dev mode
The text for transitioning to dev mode is based on a 'phy_rec'
boolean in boards.yaml (for physical recovery button vs. keyboard).
Another board now needs to use the power button for this purpose,
so I've changed this to a tristate enum (power, recovery, keyboard).

The text basically now reflects what button is used for physical
presence detection, so I've also renamed the attribute from
'phy_rec' to 'phy_pres'.

BUG=b:129471321
BRANCH=None
TEST="make" and observed that mappings reported to stdout
reflect what's in boards.yaml and that a board using the new
'power' option used the mapping for power. Verified that
power button is referenced on screen for transitioning to
dev mode.

Change-Id: I50adb6135c25731dd2804853101fd608c11f9533
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1581121
Commit-Ready: Matt Delco <delco@chromium.org>
Tested-by: Matt Delco <delco@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>