archive_image: Fix image paths
The function archive_images() expects 'files' to be relative to the
'output' directory. However, if the 'output' argument is a relative
path, when archiving localized images, the 'files' argument passed will
be paths relative to the working directory, leading to incorrect paths
in archive_images(). The 'emerge chromeos-bmpblk' command has no errors
just because the 'output' argument is a absolute path.

This patch changes the 'files' argument to paths relative to the working
directory.

BRANCH=none
BUG=none
TEST=make OUTPUT="build/nami" ARCHIVER="/usr/bin/archive" archive;
     emerge-nami chromeos-bmpblk

Change-Id: I6385783dac256df6c845e8f384bed2459314b94a
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>