images: Add right-to-left information to locales
In the menu UI, the text is no longer center-aligned. Therefore,
we need to know whether each locale should be left-aligned or
right-aligned in the firmware screen.

Add 'rtl' to indicate whether each locale is a right-to-left language.
This changes the file 'locales' to be a CSV file, with each line being
the format "code,rtl", where 'code' is the language code, and 'rtl' is
either '0' (left-to-right) or '1' (right-to-left).

Also check whether the values of rtl and hi_res are contained in locales
and raise BuildImageError on errors.

In addition, update README and fix some errors in it.

BUG=b:146399181, b:150110223
TEST=emerge-nami chromeos-bmpblk

Cq-Depend: chromium:2112318
Change-Id: I782e66b0887689de1fcfaf15a2780e75a5535b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2112272
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>