images: Use 72DPI when converting SVG to PNG files
The image scaling of rsvg-convert during processing the DPI options
may cause the rightmost or bottommost pixels of the generated PNG images
to be transparent (or filled with the specified background color). For
example, when the SVG file has width=37pt and height=26pt, the
resolution of the output PNG file will be 46x33, with the botton 2 rows
being transparent.

Since the width and height of the SVG files generated by pango-view are
in points, use the standard 72DPI for rsvg-convert to avoid the extra
image scaling. Note that the '--width' and '--height' options also imply
image scaling, but don't have the same issue.

BUG=b:146399181
TEST=MENU_UI=1 make -C images -b nami

Change-Id: Ieb13a67974261ea19d2b47eae30dfafdb624f169
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2162533
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Auto-Submit: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>