boards.yaml: Add Trogdor
Add firmware screen scaling information for Trogdor derivatives. Due to
b:157510689 we unfortunately can only define one size for the whole
family at this point. Word on the street is that we're not planning to
make anything larger than 1920x1080, so let's set it to that and have
the rest rely on downscaling.

BUG=b:157507955
TEST=none

Change-Id: Ib0f6751cce2325f53d061f974fc9a462ae530e43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2219240
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>