strings: Remove horizontal margin for font bitmaps
CL:2234871 accidentally changed the margin of characters bitmaps from
"3 0" to "3". That is, the left and right margins are increased from 0
to pango-view's default. This causes vboot_draw_dev_signed_warning() in
depthcharge to draw a text that exceeds the canvas boundary.

Therefore, change it back to "3 0" for legacy menu UI.

BUG=b:160111774
TEST=make -C strings font

Change-Id: I1b90d92a2e3e395e71a57a671cb0b4841f94bcd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2282645
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>