bmpblk: Support multiline text bitmaps
Support generation of multiline text bitmaps, by adding the option of
telling pango-view the width of the space available for the text.

Also, store the number of lines in the header of the resulting .bmp
file, allowing Depthcharge to do appropriate scaling.

BUG=b:158634754
BRANCH=puff

Signed-off-by: jbk@chromium.org
Change-Id: Ib2e7ee1dda845ea9336af0da48f1045124b01e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2303209
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>