images: Fix calculation of num_lines
Currently the number of lines of text in an SVG file is determined by
counting the <g> tags in it (CL:2303209). This works for English text,
but is incorrect for other languages such as Japanese and Arabic. Some
of the calculated num_lines is therefore larger than the correct value,
leading to unnecessarily large bitmaps stored into CBFS.

Since there is no easy way to derive the number of lines directly from
the SVG file generated by pango-view, in this patch we derive the number
by comparing the height of the generated PNG file with the one without
'--width' passed to pango-view. As a result, for menu UI, 2 PNG files
will be generated by pango-view for each localized text.

In addition, since depthcharge only supports drawing multi-line texts
for descriptions now, similar to strings/locale/build.py, we only
calculate num_lines for description strings. Other strings such as
titles and non-text images will be skipped.

BUG=b:146399181, b:158634754
TEST=emerge-puff chromeos-bmpblk

Change-Id: I1c49e91b9613a7c76592a1a3b9dd76c66a37f8d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2342518
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>