strings: Use Roboto as the default font
Use Roboto for Latin and Greek languages, and override the other
languages with 'Noto Sans' fonts.

Since the text image will be scaled to a fixed height in runtime,
changing the font potentially affects the text size rendered on the
screen. As an example, for Arabic (ar), with the same font size passed
to pango-view, the height of the generated image of language name will
decrease from 75px to 49px. As a result, when scaling to a fixed height
in depthcharge, the text size will roughly increase by 1.5 times.

BUG=b:146399181, b:160673553, b:160837806
TEST=MENU_UI=0 make -C strings
TEST=MENU_UI=1 make -C strings

Change-Id: Ia3d810e1b3ec55e02487469e3b3c647d26eb6c0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2344527
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Auto-Submit: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>