bmpblk: Determine num_lines from SVG files
With the '--width' option passed to pango-view, the number of lines in
the generated text image will be different for PNG and SVG files. Since
we are using SVG files for localized texts, the number of lines should
also be determined by comparing the heights of SVG files instead of PNG
ones. The height of an SVG image is obtained from the 'height' attribute
of the <svg> tag.

In addition, change the default value of the '--outdir' option to be the
working directory ".".

BUG=b:146399181, b:165176389
TEST=emerge-puff chromeos-bmpblk
TEST=Description text in to-norm screen is not stretched in Spanish

Change-Id: If78b1c4a7f31e0077708fe376b7194b8352a9fc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2362411
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Auto-Submit: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>