bmpblk: Adjust language order
The current language order is kind of random. Adjust it to be consistent
with the order in Chromebook Help Center, except for English (en), which
is placed at the beginning of the list.

WARNING: This CL should **never** be cherry-picked to firmware branches
which have already locked down RO, since this CL potentially changes the
index of the locale that has been saved to nvdata.

BUG=b:146399181, b:168680637
TEST=emerge-puff chromeos-bmpblk
BRANCH=none

Change-Id: If8e84579041166d8d8b3e605e8d9af3397f66923
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2418773
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Auto-Submit: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>