bmpblk: Introduce dpi option to boards.yaml
Add a dpi option to boards.yaml to control the resolution of the
generated text bitmaps. The font size passed to pango-view is calculated
from the 'height' config in format.yaml, so that the bitmap resolution
will be roughly proportional to the height as before. Therefore,
FONT_SIZE can be removed from Makefile. See go/bmpblk-scale for details.

The image generation flow is also changed in this patch. For text
bitmaps (both generic and localized bitmaps), we use pango-view to
generate PNG files directly, so no SVGs are involved anymore. As a
result, 'num_lines' is calculated by comparing the height of multi-line
and one-line PNGs, both of which are generated by the text_to_png_svg
script.

When the size of a bitmap is larger than what will be displayed at
runtime, we downscale it based on the 'screen' config in boards.yaml. To
achieve that, we revert the changes in CL:2291719 so that the config can
match the physical panel resolution. As a direct consequence, the size
of glyph bitmaps will become the same as they are rendered.

BUG=b:147123923, b:168810876
TEST=emerge-asurada chromeos-bmpblk chromeos-bootimage
TEST=emerge-puff chromeos-bmpblk chromeos-bootimage
TEST=emerge-volteer chromeos-bmpblk chromeos-bootimage
BRANCH=none

Change-Id: I9b192ae9e8eb6d8c2be150958bb763e09f41b179
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2549365
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>