bmpblk: Copy English bitmaps of untranslated strings
Currently we fill in missing entries in the translation dict with
English strings, so that before the translated strings are checked in,
screen drawing will not fail at runtime. However, when converting these
English strings to bitmaps, we are using the specified font for each
locale instead of the font for 'en', and this causes the width check
(Converter._check_text_width()) to fail.

To fix it, instead of filling missing entries in an early stage, we copy
the .bmp files for those untranslated strings from the 'en' directory.

BUG=b:177501043, b:144969091
TEST=make BOARD=asurada
BRANCH=none

Change-Id: I2f6820fa60243b53ae1a6fbec14c2a381a80d497
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2623536
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>