bmpblk: Remove right margin for URL
When '--width' option is passed to pango-view, the generated bitmap will
be as wide as the specified value, even if the text can fit in one line.
Since the dev mode URL will be used (in CL:2682057) in the error box for
*all* languages, including right-to-left ones, the large right margin of
the URL bitmap needs to be removed.

Add a new category 'error_url' for 'dev_mode_url', and keep `max_width`
null for the category.

BUG=b:144969091
TEST=make BOARD=asurada
TEST=No right margin in the generated dev_mode_url.bmp
BRANCH=none

Change-Id: I021aa90d7eb0d3f61f7deb589bce3637d1804098
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/2762485
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>