boards: add the shotzo
We merge the shotzo to the dedede OS image and the shotzo will be one
of unibuild projects under the dedede. To support the different bmpblk
settings like physical_presence_recovery, we need to have a dedicated
bmpblk setting for the shotzo. As a result, we add the shotzo to the
boards.yaml.

BUG=b:235983675
BRANCH=firmware-dedede
TEST=1. build pass
     2. flash the firmware to the device and check the recovery mode

Change-Id: I587f90aa01e2c769b6959122f6008bd8741d4dcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/3893917
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Liam Flaherty <liamflaherty@chromium.org>