Add rw_only board config attribute/key
This attribute is intended to keep the subset of firmware UI assets that
are not required during recovery mode in RW section only. This should be
used only to prevent crossing the RO section size threshold.

BUG=b:281095856
BRANCH=None
TEST=Build Myst BIOS image with some FW UI assets only in RW section.
This helped to get the RO section under 4 MB with spare 150 KB.

Change-Id: If92aefd5912755d401e179403efb84978021da37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/bmpblk/+/4519687
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@google.com>